task: moved components to views folders
